Decouple encryption from the store
Also fixes the IV being constant for multiple values causing the plaintext to always encrypt to the same ciphertext.

Right now the encryption algorithm is stored together with a key as an enum. There may be more performant ways of achieving this that can also adapt to more algorithms using traits.